+++ shub/apps/api/management/commands/cleanup_dummy.py
@@ -10,7 +10,6 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 from django.core.management.base import BaseCommand
 from shub.apps.main.models import Container
-import json
 
 
 class Command(BaseCommand):